Add support for background processes
This feature allows us to setup servers at the beginning of a cram test
and interact with it. Backgrounded processes are torn down at the end of
a cram test file. One limitation to be aware of -- backgrounded
processes will be run before any other commands no matter where they are
defined in the cram test file!